requirements.txt
uwsgi
flask
rpi.gpio
